refactored some User functions
Added to ValidationController
+ CheckPassword
+ PasswordLength
+ PasswordContainsNumber
+ PasswordContainsCapitalLetter
+ RandomString
+ CalculateAge
+ HashPassword

Functions added have been removed from user. And all references have been set to right class
